requirements.txt
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
debugpy==1.8.14
decorator==5.2.1
Django==5.1.7
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-easy-audit==1.3.7
django-environ==0.11.2
django-filter==24.3
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
environ==1.0
et_xmlfile==2.0.0
executing==2.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2024.2
pywin32==310
pyzmq==26.4.0
qrcode==8.0
razorpay==1.4.2
requests==2.32.3
service-identity==24.2.0
setuptools==78.1.0
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
